+++ .storybook/preview.js
@@ -13,10 +13,15 @@ import './drupal/drupal';
 import './drupal/once';
 
 // Import component behaviors (with JS)
-import '../source/patterns/components/accordion/accordion.js';
 import '../source/patterns/components/alert/alert.js';
 import '../source/patterns/components/carousel/carousel.js';
 import '../source/patterns/components/dropdown/dropdown.js';
+
+// Import element behaviors
+import '../source/patterns/elements/collapse/collapse.js';
+
+// Import collection behaviors
+import '../source/patterns/collections/accordion/accordion.js';
 
 // Import base pattern behaviors
 import '../source/patterns/base/animations/animations.js';